+++ MANIFEST.in
@@ -1,5 +1,3 @@
-include services/utils/requirements.txt
-include services/data/requirements.txt
 include requirements.dev.txt
 include requirements.txt
 include README.md
