+++ requirements.txt
@@ -1,4 +1,2 @@
--r services/utils/requirements.txt
--r services/data/requirements.txt
 -r services/metadata_service/requirements.txt
 -r services/migration_service/requirements.txt
