+++ services/metadata_service/api/utils.py
@@ -1,5 +1,4 @@
 import json
-import logging
 from functools import wraps
 
 import pkg_resources
@@ -34,14 +33,12 @@ def web_response(status: int, body):
                         body=json.dumps(body),
                         headers=MultiDict(
                             {"Content-Type": "application/json",
-                             "Access-Control-Allow-Origin": "*",
                              METADATA_SERVICE_HEADER: METADATA_SERVICE_VERSION}))
 
 
-def http_500(msg, id, traceback_str=get_traceback_str()):
+def http_500(msg, traceback_str=get_traceback_str()):
     # NOTE: worth considering if we want to expose tracebacks in the future in the api messages.
     body = {
-        'id': id,
         'traceback': traceback_str,
         'detail': msg,
         'status': 500,
@@ -60,14 +57,6 @@ def handle_exceptions(func):
         try:
             return await func(*args, **kwargs)
         except Exception as err:
-            # pass along an id for the error
-            err_id = getattr(err, 'id', None)
-            # either use provided traceback from subprocess, or generate trace from current process
-            err_trace = getattr(err, 'traceback_str', None) or get_traceback_str()
-            if not err_id:
-                # Log error only in case it is not a known case.
-                err_id = 'generic-error'
-                logging.error(err_trace)
-            return http_500(str(err), err_id, err_trace)
+            return http_500(str(err))
 
     return wrapper
